// app/src/api/knowledge-graph/KnowledgeGraphClient.ts
export interface Document {
  id: string;
  title: string;
  content: string;
  author: string;
  created_at: number;
}
// ...
export interface Link {
  id: string;
  source_doc_id: string;
  source_text: string;
  target_doc_id: string;
  target_text: string;
  author: string;
  created_at: number;
}